Criando edição de Contas

+++ src/repository/ContasRepository.ts
@@ -1,6 +1,7 @@
 import { createQueryBuilder, getConnection, getManager } from "typeorm";
 import Bancos from "../entity/Bancos";
 import { Contas } from "../entity/Contas";
+import { ObjetivosFinaceiros } from "../entity/ObjetivosFinaceiros";
 
 import { Usuarios } from "../entity/Usuarios";
 
@@ -76,7 +77,7 @@ export default class ContasRepository {
 
    async readConta(idUsuario: number) {
 
-     
+
       let retornoContas: Contas[]
 
       try {
@@ -84,7 +85,7 @@ export default class ContasRepository {
          const contaRepository = await createQueryBuilder("Contas")
             .leftJoinAndSelect('Contas.bancosIdFK', 'banco')
             .leftJoinAndSelect('Contas.usuariosIdFK', 'usuarios')
-         .where('usuarios.id = :id', { id: idUsuario  })
+            .where('usuarios.id = :id', { id: idUsuario })
             .getMany()
 
 
@@ -119,16 +120,38 @@ export default class ContasRepository {
    }
 
 
-   async deleteContaId(idConta: number){
+   async deleteContaId(idConta: number) {
 
-   try {
-      const contasRepository = getManager();
-     await contasRepository.delete(Contas, idConta);
-     return `ID Conta ${idConta} deletado!`
-   }catch(e){
-      return { Error: `Erro ao deletar!` , e}
+      try {
+         const contasRepository = getManager();
+         let contaVerificacao = await createQueryBuilder("ObjetivosFinaceiros")
+            .leftJoinAndSelect('ObjetivosFinaceiros.contasIdFK', 'contas')
+            .where('contas.id = :id', { id: idConta })
+            .getOne()
+
+         if (!contaVerificacao) {
+
+            contaVerificacao = await createQueryBuilder("Movimentacoes")
+               .leftJoinAndSelect('Movimentacoes.contasIdFK', 'contas')
+               .where('contas.id = :id', { id: idConta })
+               .getOne()
+
+            if (!contaVerificacao) {
+
+               await contasRepository.delete(Contas, idConta);
+               return `ID Conta ${idConta} deletado!`
+            } else {
+               return `Não foi possivel deletar , Já exite movimentações na conta!`
+            }
+         } else {
+            return `Não foi possivel deletar , Já exite Obketivos vinculads a conta na conta!`
+         }
+
+      } catch (e) {
+         return { Error: `Erro ao deletar!`, e }
+      }
+
    }
 
-}
 
 }
